After building and running with icc

+++ InfectiousDiseaseModeling/GraphHandler.cpp
@@ -6,9 +6,9 @@
 #include "GraphHandler.h"
 
 // Scan the location graph and return a map that binds every location with a vector of neighbouring locations
-std::map<size_t, std::vector<size_t>> GraphHandler::get_node_neighborhood_lookup_map(const LocationUndirectedGraph& location_graph) {
+std::map<int, std::vector<int>> GraphHandler::get_node_neighborhood_lookup_map(const LocationUndirectedGraph& location_graph) {
 	
-	std::map<size_t, std::vector<size_t>> returning_neighborhood_lookup_map;
+	std::map<int, std::vector<int>> returning_neighborhood_lookup_map;
 
 	LocationUndirectedGraph::vertex_iterator vertex_iterator_start, vertex_iterator_end; // Location node iterators
 	std::tie(vertex_iterator_start, vertex_iterator_end) = vertices(location_graph); // Tie iterators with the current graph
@@ -17,7 +17,7 @@ std::map<size_t, std::vector<size_t>> GraphHandler::get_node_neighborhood_lookup
 	for (; vertex_iterator_start != vertex_iterator_end; ++vertex_iterator_start) {
 
 		tie(neighbour_iterator_start, neighbour_iterator_end) = adjacent_vertices(*vertex_iterator_start, location_graph); // Tie adjacent/neighbouring location nodes
-		std::vector<size_t> current_neighborhood;
+		std::vector<int> current_neighborhood;
 		for (; neighbour_iterator_start != neighbour_iterator_end; ++neighbour_iterator_start)
 			current_neighborhood.push_back(*neighbour_iterator_start); // Add the current neighbour
 
@@ -27,10 +27,10 @@ std::map<size_t, std::vector<size_t>> GraphHandler::get_node_neighborhood_lookup
 	return returning_neighborhood_lookup_map;
 }
 
-// TODO: return "location size_t" vector from graph vertices (by tie of the graph iterators). Must be a distinct list
+// TODO: return "location int" vector from graph vertices (by tie of the graph iterators). Must be a distinct list
 
 // Generate a vector of individuals and assign a random location within the requested ranges
-std::vector<Individual> GraphHandler::get_random_individuals(size_t individual_count, size_t location_count) {
+std::vector<Individual> GraphHandler::get_random_individuals(int individual_count, int location_count) {
 
 	// Generate a population of healthy individuals
 	std::vector<Individual> individuals(individual_count, Individual());
@@ -38,10 +38,10 @@ std::vector<Individual> GraphHandler::get_random_individuals(size_t individual_c
 	// Randomly assign locations to individuals in the population	
 	std::random_device random_device;
 	std::mt19937 mersenne_twister_engine(random_device());
-	std::uniform_int_distribution<> uniform_int_distribution(0, static_cast<int>(location_count));
+	std::uniform_int_distribution<> uniform_int_distribution(0, location_count);
 
 	for (Individual& current_individual : individuals) {
-		current_individual.set_location(static_cast<size_t>(uniform_int_distribution(mersenne_twister_engine))); // Assign the random location
+		current_individual.set_location(uniform_int_distribution(mersenne_twister_engine)); // Assign the random location
 	}
 	return individuals;
 }
@@ -64,15 +64,15 @@ LocationUndirectedGraph GraphHandler::get_location_undirected_graph_from_file(st
 	vector<string> string_vector;
 	string current_line;
 
-	map<size_t, uint32_t> map_location_to_index;
-	uint32_t current_index = 0;
+	map<int, int> map_location_to_index;
+	int current_index = 0;
 
 	while (getline(input_file_stream, current_line)) {
 		Tokenizer tok(current_line);
 		string_vector.assign(tok.begin(), tok.end());
 
-		size_t first_edge = stoull(string_vector[0]);
-		size_t second_edge = stoull(string_vector[1]);
+		int first_edge = stoull(string_vector[0]);
+		int second_edge = stoull(string_vector[1]);
 
 		if (map_location_to_index.find(first_edge) == map_location_to_index.end()) {
 			// not found
@@ -131,14 +131,14 @@ void GraphHandler::save_undirected_graph_to_graphviz_file(std::string filename, 
 
 	std::ofstream dotfile(filename.c_str());
 	boost::write_graphviz(dotfile, location_graph,
-		boost::make_label_writer(&NameVec[0]) // TODO: bind id with the map lookup size_t to string
+		boost::make_label_writer(&NameVec[0]) // TODO: bind id with the map lookup int to string
 		);
 
 	dotfile.close();
 }
 
 // Save the hit and infected counts for each epoch into a csv file, to disk
-void GraphHandler::save_epoch_statistics_to_csv(std::string filename, const std::vector<std::tuple<size_t, size_t>> epoch_statistics){
+void GraphHandler::save_epoch_statistics_to_csv(std::string filename, const std::vector<std::tuple<int, int>> epoch_statistics){
 
 	std::ofstream output_csv;
 	output_csv.open(std::string(filename));
@@ -147,24 +147,24 @@ void GraphHandler::save_epoch_statistics_to_csv(std::string filename, const std:
 	output_csv << "epoch,hitcount,infectedcount" << std::endl;
 	
 	// Write a line for each epoch
-	for (size_t epoch_index = 0; epoch_index != epoch_statistics.size(); ++epoch_index)
+	for (int epoch_index = 0; epoch_index != epoch_statistics.size(); ++epoch_index)
 		output_csv << epoch_index << "," << get<0>(epoch_statistics[epoch_index]) << "," << get<1>(epoch_statistics[epoch_index]) << std::endl;
 
 	output_csv.close();
 }
 
 // Show the Hit percentage (fraction of the total population that got infected), epidemic peak percentage and the epoch of the epidemic peak
-void GraphHandler::show_epidemic_results(size_t population_count, const std::vector<std::tuple<size_t, size_t>> epoch_statistics) {
+void GraphHandler::show_epidemic_results(int population_count, const std::vector<std::tuple<int, int>> epoch_statistics) {
 	
 	// Fraction of the population that got infected
-	size_t hit_count = get<0>(epoch_statistics[epoch_statistics.size() - 1]);
+	int hit_count = get<0>(epoch_statistics[epoch_statistics.size() - 1]);
 
 	// Epidemic peak %
-	size_t epidemic_peak_size = 0;
-	size_t epidemic_peak_epoch = 0;
+	int epidemic_peak_size = 0;
+	int epidemic_peak_epoch = 0;
 
 	// Write a line for each epoch
-	for (size_t epoch_index = 0; epoch_index != epoch_statistics.size(); ++epoch_index) {
+	for (int epoch_index = 0; epoch_index != epoch_statistics.size(); ++epoch_index) {
 		if (get<1>(epoch_statistics[epoch_index]) > epidemic_peak_size) {
 			epidemic_peak_size = get<1>(epoch_statistics[epoch_index]);
 			epidemic_peak_epoch = epoch_index;
@@ -172,7 +172,7 @@ void GraphHandler::show_epidemic_results(size_t population_count, const std::vec
 	}
 
 	std::cout << std::endl << "-- Epidemic Results --" << std::endl;
-	std::cout << "Hit: " << static_cast<float>(hit_count) / static_cast<float>(population_count) << " %"<< std::endl;
-	std::cout << "Epidemic Peak:" << static_cast<float>(epidemic_peak_size) / static_cast<float>(population_count) << " %" << std::endl;
+	std::cout << "Hit: " << static_cast<double>(hit_count) / static_cast<double>(population_count) << " %"<< std::endl;
+	std::cout << "Epidemic Peak:" << static_cast<double>(epidemic_peak_size) / static_cast<double>(population_count) << " %" << std::endl;
 	std::cout << "Epidemic Peak Epoch: " << epidemic_peak_epoch << std::endl;
 }
